Fix a few ambiguous overloaded calls
Noticed this when exploring another optimisation

By specification exactly which overloaded function to use
we save some memory

Seems to have something to do with doing arithmetic in the function argument

Change in memory:
Flash: -156 bytes
SRAM: 0 bytes

+++ Firmware/Marlin_main.cpp
@@ -3456,7 +3456,7 @@ static void mmu_M600_unload_filament() {
     lcd_clear();
     lcd_puts_at_P(0, 1, _T(MSG_UNLOADING_FILAMENT));
     lcd_print(' ');
-    lcd_print(MMU2::mmu2.get_current_tool() + 1);
+    lcd_print(uint8_t(MMU2::mmu2.get_current_tool() + 1));
 
     // unload just current filament for multimaterial printers (used also in M702)
     MMU2::mmu2.unload();
